Add assertion and casting utilities
Because of the exact implementation of the Ukkonen's algorithm, there are
a number of places where the indices can be negative, forcing us to cast
back and forth between size_t and ssize_t. These utilities deal with
those cases to call out at run-time if we end up doing something horrible.